test: test framework and tests for initial version
This includes unit tests and integration tests for the initial
working version of the Data Downloader. It includes unit tests
for some of the functionality, and integration tests covering
the core functions of the Data Downloader. Integration tests
are run against a docker compose setup which spins up
an Azurite storage emulator and a Mockoon server.